+++ discord/flags.py
@@ -15,7 +15,7 @@ class ActivityFlags(IntFlag):
 
 
 @unique
-class ApplicationFlag(IntFlag):
+class ApplicationFlags(IntFlag):
   APPLICATION_AUTO_MODERATION_RULE_CREATE_BADGE: int = 1 << 6
   """Indicates if an app uses the Auto Moderation API"""
 
@@ -48,7 +48,7 @@ class ApplicationFlag(IntFlag):
 
 
 @unique
-class ChannelFlag(IntFlag):
+class ChannelFlags(IntFlag):
    PINNED: int = 1 << 1
    """This thread is pinned to the top of its parent :attr:`~discord.enums.ChannelType.GUILD_FORUM` or :attr:`~discord.enums.ChannelType.GUILD_MEDIA` channel"""
 
@@ -63,7 +63,7 @@ class ChannelFlag(IntFlag):
 
 
 @unique
-class GatewayIntent(IntFlag):
+class GatewayIntents(IntFlag):
   """Represents a set of Gateway intents.
 
   .. note::
@@ -232,7 +232,7 @@ class GuildInviteFlags(IntFlag):
 
 
 @unique
-class GuildMemberFlag(IntFlag):
+class GuildMemberFlags(IntFlag):
   DID_REJOIN: int = 1 << 0
   """Member has left and rejoined the guild.
 
@@ -298,7 +298,7 @@ class GuildMemberFlag(IntFlag):
 
 
 @unique
-class PermissionFlag(IntFlag):
+class PermissionFlags(IntFlag):
   """Represents a set of permission flags"""
 
   CREATE_INSTANT_INVITE: int = 1 << 0
@@ -503,13 +503,13 @@ class PermissionFlag(IntFlag):
 
 
 @unique
-class RoleFlag(IntFlag):
+class RoleFlags(IntFlag):
    IN_PROMPT: int = 1 << 0
    """Role can be selected by members in an onboarding prompt"""
 
 
 @unique
-class SystemChannelFlag(IntFlag):
+class SystemChannelFlags(IntFlag):
    SUPPRESS_JOIN_NOTIFICATIONS: int = 1 << 0
    """Suppress member join notifications"""
 
@@ -530,7 +530,7 @@ class SystemChannelFlag(IntFlag):
 
 
 @unique
-class UserFlag(IntFlag):
+class UserFlags(IntFlag):
    STAFF: int = 1 << 0
    """Discord Employee"""