+++ discord/objects/guild.py
@@ -1,6 +1,6 @@
 from .._dataclass import dataclass
 from ..enums import DefaultMessageNotificationLevel, ExplicitContentFilterLevel, GuildAgeRestrictionLevel, GuildFeature, Locale, MFALevel, PremiumTier, VerificationLevel
-from ..flags import PermissionFlag, SystemChannelFlag
+from ..flags import PermissionFlags, SystemChannelFlags
 from ..snowflake import Snowflake
 from ..utils import Nullable, Optional
 from .emoji import Emoji
@@ -93,7 +93,7 @@ class Guild:
      This field is only sent when using the :meth:`~discord._http.HTTP.get_current_user_guilds` endpoint and are relative to the requested user.
   """
 
-  permissions: Optional[PermissionFlag]
+  permissions: Optional[PermissionFlags]
   """Total permissiosn for the user in the guild (excludes overwrites and implicit permissions).
 
   .. note::
@@ -137,7 +137,7 @@ class Guild:
   stickers: Optional[list[Sticker]]
   """Custom guild stickers"""
 
-  system_channel_flags: SystemChannelFlag
+  system_channel_flags: SystemChannelFlags
   """System channel flags"""
 
   system_channel_id: Nullable[Snowflake]