+++ discord/objects/guild_member.py
@@ -1,5 +1,5 @@
 from .._dataclass import dataclass
-from ..flags import GuildMemberFlag, PermissionFlag
+from ..flags import GuildMemberFlags, PermissionFlags
 from ..snowflake import Snowflake
 from ..utils import ISO8601Timestamp, Nullable, Optional
 from .avatar_decoration_data import AvatarDecorationData
@@ -38,7 +38,7 @@ class GuildMember:
   deaf: bool
   """Whether the user is deafened in voice channels."""
 
-  flags: GuildMemberFlag
+  flags: GuildMemberFlags
   """Guild member flags represented as a bit set, defaults to ``0``."""
 
   joined_at: Optional[ISO8601Timestamp]
@@ -53,7 +53,7 @@ class GuildMember:
   pending: Optional[bool]
   """Whether the user has not yet passed the guild's Membership Screening requirements."""
 
-  permissions: Optional[PermissionFlag]
+  permissions: Optional[PermissionFlags]
   """Total permissions of the member in the channel, including overwrites, returned when in the interaction object."""
 
   premium_since: Optional[Nullable[ISO8601Timestamp]]