+++ discord/objects/user.py
@@ -1,6 +1,6 @@
 from .._dataclass import dataclass
 from ..enums import PremiumType
-from ..flags import UserFlag
+from ..flags import UserFlags
 from ..snowflake import Snowflake
 from ..utils import Nullable, Optional
 from .avatar_decoration_data import AvatarDecorationData
@@ -36,7 +36,7 @@ class User:
   email: Optional[Nullable[str]]
   """The user's email"""
   
-  flags: Optional[UserFlag]
+  flags: Optional[UserFlags]
   """The flags on a user's account"""
   
   global_name: Nullable[str]
@@ -57,7 +57,7 @@ class User:
   primary_guild: Optional[Nullable[UserPrimaryGuild]]
   """The user's primary guild"""
 
-  public_flags: Optional[UserFlag]
+  public_flags: Optional[UserFlags]
   """The public flags on a user's account"""
 
   system: Optional[bool]